allow StringDiagrams which do not contain Atom
Atom is useful when the underlying 'a' doesn't itself have a notion of
input and output size, and when these sizes need to be computed. This
leaves a number of circumstances (when the underlying 'a' already has
its own notion of size, and when we're doing a computation which doesn't
require knowing those sizes, e.g. when interpreting the string diagram)
in which that extra information is not needed.