check that the normalized diagram is equivalent
I was previously planning to do this generating random boolean
functions, but I've realized that a single injective function will catch
all the divergences while a random boolean function will only catch a
few.

in the process of testing this, I found and fixed a bug in the
Widen's Interpret instance.